--- core/agent_harness/session/integration/__init__.py
@@ -1,1 +1,0 @@
-"""Per-session integration resolution, its cache, and the warmed IntegrationState."""
