+++ tools/investigation/stages/resolve_integrations/node.py
@@ -4,8 +4,8 @@ from __future__ import annotations
 
 from typing import Any
 
-from core.agent_harness.session.integration.state import resolve_integrations_with_metadata
 from core.state import InvestigationState
+from platform.harness_ports import resolve_integrations_with_metadata
 from platform.observability import get_progress_tracker as get_tracker
